+++ app/src/main/java/com/ex/unamic/pec/CategoryFragment.java
@@ -124,60 +124,7 @@ public class CategoryFragment extends Fragment {
     private void getCategory() {
         categoryModels = dataAdapter.getCategoriesByUser(userId);
 
-        //Set default data
-        if (categoryModels.isEmpty()) {
 
-            //Set default data
-
-            long housingId = dataAdapter.addCategory(new CategoryModel(0, "Housing", userId));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Rent/Mortgage"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Gas"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Landline Phone"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Internet"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Mobile Phone"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Water"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Electricity"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Insurance"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Health Insurance"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, housingId, "Credit Cards"));
-
-            // daily
-            long dailyId = dataAdapter.addCategory(new CategoryModel(0, "Daily", userId));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Food"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Hair Cut"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Medication"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Contact Lenses"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Tea/Coffee"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Clothing"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Dry Cleaning"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Cosmetics"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Tuition"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, dailyId, "Sweets / Candy"));
-
-            long transportId = dataAdapter.addCategory(new CategoryModel(0, "Transport", userId));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, transportId, "Car"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, transportId, "Car Insurance"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, transportId, "Bus Pass"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, transportId, "Subway Pass"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, transportId, "Fuel"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, transportId, "Car Tax"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, transportId, "Rail Card"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, transportId, "Car Repairs"));
-
-            long funId = dataAdapter.addCategory(new CategoryModel(0, "Fun", userId));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "Cable"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "Cinema"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "Club"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "DVD Purchases"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "DVD Rentals"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "Eating Out"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "Gym"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "Magazines"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "Music Lessons"));
-            dataAdapter.addSubCategory(new SubCategoryModel(0, funId, "Sporting Events"));
-
-            categoryModels = dataAdapter.getCategoriesByUser(userId);
-        }
 
         Resources res = getResources();
 
@@ -195,7 +142,7 @@ public class CategoryFragment extends Fragment {
                     @Override
                     public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                         Object listItem = lvCategory.getItemAtPosition(position);
-                        if(categoryModels.size() > 0) {
+                        if (categoryModels.size() > 0) {
                             selectedCategory = (CategoryModel) listItem;
 
                             etCategory.setText(selectedCategory.getCategory());
@@ -213,15 +160,14 @@ public class CategoryFragment extends Fragment {
 
         );
         lvCategory.setOnItemLongClickListener(
-                new AdapterView.OnItemLongClickListener()
-                {
+                new AdapterView.OnItemLongClickListener() {
 
                     @Override
                     public boolean onItemLongClick(AdapterView<?> adapterView, View view, int i, long l) {
                         Integer index = i;
                         Log.d("Item long click", index.toString());
 
-                        if(categoryModels.size() > 0) {
+                        if (categoryModels.size() > 0) {
                             Object listItem = lvCategory.getItemAtPosition(i);
                             detailsCategory = (CategoryModel) listItem;
                         }
@@ -259,7 +205,7 @@ public class CategoryFragment extends Fragment {
                         categoryModels.add(selectedCategory);
                     }
                 } else {
-                    UtilsUI.showMessage(view, getResources(), "Your category not be blank.", true);
+                    UtilsUI.showMessage(view, getResources(), "Your category not be blank.", true, false);
                 }
 
                 updateCategoryBehavior();
@@ -278,14 +224,18 @@ public class CategoryFragment extends Fragment {
     }
 
     private void deleteCategory() {
-        if (selectedCategory == detailsCategory) {
-            updateCategoryBehavior();
-        }
-        dataAdapter.deleteCategory(detailsCategory.getId());
-        categoryModels.remove(detailsCategory);
-        detailsCategory = null;
+        if (deleteValidation()) {
+            if (selectedCategory == detailsCategory) {
+                updateCategoryBehavior();
+            }
+            dataAdapter.deleteCategory(detailsCategory.getId());
+            categoryModels.remove(detailsCategory);
+            detailsCategory = null;
 
-        categoryAdapter.notifyDataSetChanged();
+            categoryAdapter.notifyDataSetChanged();
+        } else {
+            UtilsUI.showMessage(thisView, getResources(), "There are sub categories in this category.", true, false);
+        }
     }
 
     private boolean validation(CategoryModel model) {
@@ -296,5 +246,14 @@ public class CategoryFragment extends Fragment {
         return state;
     }
 
+    private boolean deleteValidation() {
 
+        if (detailsCategory != null) {
+            List<SubCategoryModel> subCategories = dataAdapter.getSubCategoriesByCategory(detailsCategory.getId());
+            if (subCategories.size() > 0) {
+                return false;
+            }
+        }
+        return true;
+    }
 }
